Convert JS files to TS
