+++ .npmignore
@@ -2,13 +2,12 @@ src
 examples
 .github
 .babelrc
-.eslintrc.js
+.eslintrc
 .eslintignore
 .gitignore
 .prettierrc
 .npmrc
 tsdx.config.js
-webpack.config.js
 typedoc.json
 tsconfig.json
 MIGRATION.md
