+++ package.json
@@ -18,8 +18,8 @@
     },
     "scripts": {
         "prebuild": "npm run docs",
-        "build": "tsdx build --entry ./src/index.js",
-        "watch": "tsdx watch --transpileOnly --entry ./src/index.js",
+        "build": "tsdx build --entry ./src/index.ts",
+        "watch": "tsdx watch --transpileOnly --entry ./src/index.ts",
         "docs": "rimraf typedoc.json && typedoc --tsconfig",
         "lint": "eslint . --ext .js,.jsx,.ts,.tsx",
         "test": "echo todo-test",
@@ -38,13 +38,9 @@
         "@types/react-dom": "^18.0.3",
         "@typescript-eslint/eslint-plugin": "^5.28.0",
         "@typescript-eslint/parser": "^5.28.0",
-        "babel-eslint": "^10.1.0",
         "babel-plugin-optimize-clsx": "^2.6.1",
         "eslint": "^7.7.0",
-        "eslint-config-airbnb": "~18.2.0",
         "eslint-config-prettier": "^8.5.0",
-        "eslint-plugin-import": "~2.22.0",
-        "eslint-plugin-jsx-a11y": "^6.3.1",
         "eslint-plugin-react": "^7.30.0",
         "np": "^7.6.1",
         "prettier": "^2.7.1",
